requirements.txt
# Python Dependencies for EPON Network Agent ADK

# ========================================
# Core Framework
# ========================================
# Google ADK and Gemini
google-genai>=0.2.0
google-adk>=0.1.0

# ========================================
# Web Framework
# ========================================
Flask>=2.3.0
Werkzeug>=2.3.0

# ========================================
# Data Processing
# ========================================
pandas>=2.0.0
numpy>=1.24.0

# ========================================
# Visualization
# ========================================
matplotlib>=3.7.0
seaborn>=0.12.0

# ========================================
# Async & Utilities
# ========================================
asyncio>=3.4.3
typing-extensions>=4.5.0

# ========================================
# Testing (Optional)
# ========================================
pytest>=7.3.0
pytest-asyncio>=0.21.0

# ========================================
# Development Tools (Optional)
# ========================================
# Uncomment for development
# black>=23.3.0
# flake8>=6.0.0
# mypy>=1.3.0
